requirements.txt
tensorflow==2.7
tensorflow-addons
sklearn
tqdm
pandas
biopython
pillow
matplotlib
snakemake
pyyaml